+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Changed
+- Decrease the delay from typing a letter to search start to 500ms
 
 ## [0.3.0] - 2021-08-09
 ### Changed